+++ resources/views/posts/_card.blade.php
@@ -211,6 +211,12 @@
                                 <x-icon name="quote" />
                                 {{ __('openbook.actions.announce_quote') }}
                             </a>
+                            @unless ($post->isDirectMessage())
+                                <a href="{{ route('posts.share_to_user', $post) }}" class="ob-post__menu-item" role="menuitem">
+                                    <x-icon name="message" />
+                                    {{ __('openbook.actions.announce_share_user') }}
+                                </a>
+                            @endunless
                         </div>
                     </details>
                 @else